Fixed #16 by making XMLParsing restart its lists in its associated parsing functions (so that it can't accidentally keep appending if the user continues calling `parse`, for example).  To limit further the chance someone is inclined to do that, I renamed those internal methods to prefixes with __ and refactored where appropriate so that creating the XMLParsing object is enough to run the process.
Fixed #34 -- the issue was in how we were tracking references.  Also noticed we had no way to track multiple named variables in a given block definition, which is possible.  So I pre-emptively "fixed" that issue.  Switched from the namedtuple to a class since the former is immutable and we need to be able to set 'type' on the variable once it's referenced from a block.  That led to moving the XML block interpretation into the class.

Fixed #40, indentions, by refactoring the whitespace in the file to use one tab width (was 2 spaces and 4 spaces in various places).